Name of create_mac_bundle.py was not consistent with reference to it "createMacBundle" in packageMacVersion.py.  I chose the path of the least number of changes, and just renamed create_mac_bundle.py to createMacBundle.py for now, although I'm happy to discuss changes to file naming conventions away from lowerCamelCase.
createMacBundle.py contained the following line:

mydataVersionNumberModulePath = \
    os.path.dirname(pkgutil.get_loader("mydata.MyDataVersionNumber").filename)

which didn't work for me, because my "mydata" subdirectory was not a
Python module (because it didn't contain an "__init__.py").  @grischa,
maybe you created an "__init__.py" but forgot to add it to the Git repo?
I also noticed that this "mydataVersionNumberModulePath" is not actually
used anywhere in createMacBundle.py, so I removed it.

However, I do like the idea of being able to "import mydata" from the
top directory, so I have created an "__init__.py" and added it to the
repository.

Also, the code I added to createMacBundle.py to copy "favicon.ico" into
the app bundle for use in the task bar icon didn't work correctly.
MyDataTaskBarIcon.py was expecting the icon to be in
"MyData.App/Contents/Resources/media/", which it wasn't.  This has
been fixed now.