No longer using "folder_type" in FolderModel.  That was designed to allow users to drag and drop and folder into MyData and then specify that it represents a dataset or an experiment etc.  Now all folders represented by FolderModel are mapped to datasets.
Introduced userFolderName attribute, which records the user folder name
(username or email address or full name) which will be added to the experiment
metadata for default experiments created for that user.  Previously MyData
attempted to resolve the user folder name to a MyTardis user record, which was
called "owner" in MyData's code, referring the researcher who owns the data
within that user folder.  Then MyData always used the MyTardis username of
that "owner" when annotating metadata for that user's default experiment on
the current instrument.

But now MyData simply uses the user folder name in the
experiment metadata, which could be a username or email address or (in the
case of the "User Group\Instrument\Full Name\Dataset" folder structure), the
user folder name experiment metadata can be the user's full name.