Importing openSSH singleton from openssh module to allow access to GetSshControlMasterPool.
SSH ControlMaster processes are no longer associated with UploadModels or VerificationModels, so there is no need to run CleanUp on self.verificationsModel.

Now that we're using a pool for SSH ControlMaster processes, we need to clean it up (shut it down) in ShutDownUploadThreads.

Now that we're using a pool for SSH ControlMaster processes, we don't need to terminate SSH ControlMaster processes within verification threads.