This essentially rolls back this commit: 741232a7bb2338968340bb588941fe909f3f9778 which behaved badly for a large dataset (7,000 datafiles) and resulted in a series of duplicate key errors while MyData attempted to create datafile records which had already been created.
The aim of that commit was to speed up datafile lookups.  This commit
still adds a small speed-up, in that we are using
/api/v1/mydata_dataset_file/ instead of /api/v1/dataset_file/, so
we can ensure that the datafile lookup in the "mydata" tardis app's
API doesn't return unnecessary stuff like datafile parameters when
we just want to check for the datafile's existence.