Removing CheckIfAlreadyRunning, because of issues on Mac OS X where sometimes the mutex didn't automatically get deleted when MyData exited leading to confusion when MyData tried to launch the next time.
The single-instance-checking functionality was important when
prototyping a MyData daemon (a version of MyData which could run
as a service / daemon without any user logged in).  If we ever
properly implement the MyData-as-a-service functionality, we
will again look at ways to ensure that only one instance of the
service is running.