Fix for ipconfig parsing bug on Windows.  Sometimes "(Tentative)" appears after an IP address in ipconfig output and MyData was trying to INSERT the IP address plus the "(Tentative)" string into a database field only big enough for an IP address.
When there are multiple network interfaces, e.g. one for Internet
and one for a data-collection instrument, MyData tries to guess
which is the interface used to connect to MyTardis, so that it
can detect connectivity problems early rather than throwing an
ugly exception.  Some instrument PCs are set up with interfaces
named "Internet" and "Instrument", instead of using the more
common "Ethernet", "Local Area Connection" or
"Local Area Connection2" interface names.  In this case, MyData
was incorrectly reporting an Internet connectivity problem
when the instrument connection was inactive.