The cache lookup for each datafile is now done before a row is added to the Verifications view, and in the event of a cache hit, no row is added to the Verifications view, to avoid attempting to update the GUI too frequently.
The verifications (file lookups) view has been extended with a new panel at the bottom
showing how many files have been found in the local cache.

The CountCompletedUploadsAndVerifications method is now run on a timer, instead of in
response to events like completing a file lookup.  This is because when files are found
in the cache, we need to be careful to avoid triggering GUI updates too quickly.

We also have a timer for updating the summary of how many datafiles were found in the cache.