Switch from SCP to SFTP (and from subprocesses to Paramiko)
According to the OpenSSH 8.0 release notes, the scp protocol is
"outdated, inflexible and not readily fixed".

Keeping an SFTP channel open instead of running an SCP subprocess
for each upload reduces the number of SSH connections MyData
makes which is more efficient, particularly for small files.

Unlike the SFTP protocol, the SCP protocol doesn't have mkdir or
chmod built into it, so it was necessary to allow SSH connections
on the SCP server.  By using only the SFTP protocol, the SFTP
server can be configured to only allow SFTP connections, improving
security.